Renamed class to Flatten
- Renamed class to Flatten
- Fixed a bug of index
