+++ sympy/tensor/array/ndim_array.py
@@ -366,33 +366,33 @@ class NDimArray(object):
         return f(self._loop_size, self.shape, 0, self._loop_size)
 
     def __add__(self, other):
-        from sympy.tensor.array.nditer import nditer
+        from sympy.tensor.array.arrayop import Flatten
 
         if not isinstance(other, NDimArray):
             raise TypeError(str(other))
 
         if self.shape != other.shape:
             raise ValueError("array shape mismatch")
-        result_list = [i+j for i,j in zip(nditer(self), nditer(other))]
+        result_list = [i+j for i,j in zip(Flatten(self), Flatten(other))]
 
         return type(self)(result_list, self.shape)
 
     def __sub__(self, other):
-        from sympy.tensor.array.nditer import nditer
+        from sympy.tensor.array.arrayop import Flatten
 
         if not isinstance(other, NDimArray):
             raise TypeError(str(other))
 
         if self.shape != other.shape:
             raise ValueError("array shape mismatch")
-        result_list = [i-j for i,j in zip(nditer(self), nditer(other))]
+        result_list = [i-j for i,j in zip(Flatten(self), Flatten(other))]
 
         return type(self)(result_list, self.shape)
 
     def __mul__(self, other):
         from sympy.matrices.matrices import MatrixBase
         from sympy.tensor.array import SparseNDimArray
-        from sympy.tensor.array.nditer import nditer
+        from sympy.tensor.array.arrayop import Flatten
 
         if isinstance(other, (Iterable, NDimArray, MatrixBase)):
             raise ValueError("scalar expected, use tensorproduct(...) for tensorial product")
@@ -403,13 +403,13 @@ class NDimArray(object):
                 return type(self)({}, self.shape)
             return type(self)({k: other*v for (k, v) in self._sparse_array.items()}, self.shape)
 
-        result_list = [i*other for i in nditer(self)]
+        result_list = [i*other for i in Flatten(self)]
         return type(self)(result_list, self.shape)
 
     def __rmul__(self, other):
         from sympy.matrices.matrices import MatrixBase
         from sympy.tensor.array import SparseNDimArray
-        from sympy.tensor.array.nditer import nditer
+        from sympy.tensor.array.arrayop import Flatten
 
         if isinstance(other, (Iterable, NDimArray, MatrixBase)):
             raise ValueError("scalar expected, use tensorproduct(...) for tensorial product")
@@ -420,13 +420,13 @@ class NDimArray(object):
                 return type(self)({}, self.shape)
             return type(self)({k: other*v for (k, v) in self._sparse_array.items()}, self.shape)
 
-        result_list = [other*i for i in nditer(self)]
+        result_list = [other*i for i in Flatten(self)]
         return type(self)(result_list, self.shape)
 
     def __div__(self, other):
         from sympy.matrices.matrices import MatrixBase
         from sympy.tensor.array import SparseNDimArray
-        from sympy.tensor.array.nditer import nditer
+        from sympy.tensor.array.arrayop import Flatten
 
         if isinstance(other, (Iterable, NDimArray, MatrixBase)):
             raise ValueError("scalar expected")
@@ -435,7 +435,7 @@ class NDimArray(object):
         if isinstance(self, SparseNDimArray) and other != S.Zero:
             return type(self)({k: v/other for (k, v) in self._sparse_array.items()}, self.shape)
 
-        result_list = [i/other for i in nditer(self)]
+        result_list = [i/other for i in Flatten(self)]
         return type(self)(result_list, self.shape)
 
     def __rdiv__(self, other):
@@ -443,12 +443,12 @@ class NDimArray(object):
 
     def __neg__(self):
         from sympy.tensor.array import SparseNDimArray
-        from sympy.tensor.array.nditer import nditer
+        from sympy.tensor.array.arrayop import Flatten
 
         if isinstance(self, SparseNDimArray):
             return type(self)({k: -v for (k, v) in self._sparse_array.items()}, self.shape)
 
-        result_list = [-i for i in nditer(self)]
+        result_list = [-i for i in Flatten(self)]
         return type(self)(result_list, self.shape)
 
     def __eq__(self, other):
@@ -500,8 +500,8 @@ class NDimArray(object):
         return self._eval_transpose()
 
     def _eval_conjugate(self):
-        from sympy.tensor.array.nditer import nditer
-        return self.func([i.conjugate() for i in nditer(self)], self.shape)
+        from sympy.tensor.array.arrayop import Flatten
+        return self.func([i.conjugate() for i in Flatten(self)], self.shape)
 
     def conjugate(self):
         return self._eval_conjugate()
